# api.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "handsfree/handsfree.proto",
    "version": "version not set"
  },
...
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/handsfree": {
      "post": {
        "operationId": "HandsfreeService_Send",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/handsfreeControlPacket"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/handsfreeControlPacket"
            }
          }
        ],
        "tags": [
          "HandsfreeService"
        ]
      }
    },
    "/handsfree/connection": {
      "get": {
        "operationId": "HandsfreeService_Connect",
        "responses": {
          "200": {
            "description": "A successful response.(streaming responses)",
            "schema": {
              "type": "object",
              "properties": {
                "result": {
                  "$ref": "#/definitions/handsfreeControlPacket"
                },
                "error": {
                  "$ref": "#/definitions/rpcStatus"
                }
              },
              "title": "Stream result of handsfreeControlPacket"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "appId",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "HandsfreeService"
        ]
      }
    },
...
  },
  "definitions": {
...
    "handsfreeCode": {
      "type": "string",
      "enum": [
        "UNKNOWN",
        "AUTH",
        "DATA",
        "SUCCESS",
        "ERROR"
      ],
      "default": "UNKNOWN"
    },
    "handsfreeControlPacket": {
      "type": "object",
      "properties": {
        "code": {
          "$ref": "#/definitions/handsfreeCode"
        },
        "payload": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
...
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
...
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
...
  }
}